LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/send-network-error-async-events.sub-expected.txt

CONSOLE MESSAGE: Error resolving “nonexistent.web-platform.test”: No address associated with hostname
CONSOLE MESSAGE: XMLHttpRequest cannot load http://nonexistent.web-platform.test:8800/ due to access control checks.

PASS XMLHttpRequest: The send() method: Fire a progress event named error when Network error happens (synchronous flag is unset)
